pyproject.toml
[project]
name = "travel-recommendation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "Django>=5.2,<5.3", 
    "numpy>=2.2.4,<2.3.0",
    "pytest>=8.3.5,<8.4.0",
    "ruff>=0.11.4,<0.12.0",
    "pandas>=2.2.3,<2.3.0",
    "polars>=1.27.1,<1.28.0",
    "requests>=2.32.3,<2.33.0",
    "python-dotenv>=1.1.0,<1.2.0",
    "requests-cache>=1.2.1,<1.3.0",
    "retry-requests>=2.0.0,<2.1.0",
    "openmeteo-requests>=1.4.0,<1.5.0",
    "djangorestframework>=3.16.0,<3.17.0",
    "djangorestframework-simplejwt>=5.5.0,<5.6.0",
]

[tool.uv]
managed = false
